requirements.txt
img2pdf
Pillow 
flask
pdfplumber
pytesseract
